news_bot/requirements.txt
aiocsv==1.2.2
aiofiles==0.8.0
aiogram==2.21
aiohttp==3.8.1
aiomysql==0.1.1
aiosignal==1.2.0
altgraph==0.17.2
appdirs==1.4.4
APScheduler==3.10.0
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
asynctest==0.13.0
attrs==22.1.0
Automat==20.2.0
Babel==2.9.1
backports.zoneinfo==0.2.1
bcrypt==4.0.1
beautifulsoup4==4.11.1
bitarray==2.7.3
brotlipy==0.7.0
bs4==0.0.1
cached-property==1.5.2
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
constantly==15.1.0
cryptography==39.0.0
cssselect==1.1.0
cytoolz==0.12.1
defusedxml==0.7.1
et-xmlfile==1.1.0
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
fake-useragent==0.1.11
Flask==2.1.2
Flask-Login==0.6.1
frozenlist==1.3.0
future==0.18.2
greenlet==2.0.1
h11==0.13.0
hexbytes==0.3.0
idna==3.4
importlib-metadata==4.12.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
json5==0.9.8
jsonschema==4.17.3
lru-dict==1.1.8
lxml==4.9.1
macholib==1.16.2
magic-filter==1.0.9
MarkupSafe==2.1.1
mnemonic==0.20
multidict==6.0.2
numpy==1.21.6
opencv-python==4.7.0.68
openpyxl==3.1.1
outcome==1.2.0
pandas==1.3.5
parse==1.19.0
parsimonious==0.9.0
pefile==2022.5.30
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pycurl==7.43.0.5
PyDispatcher==2.0.5
pyee==8.2.2
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
PyMySQL==1.0.2
pyOpenSSL==23.0.0
pyppeteer==1.0.2
pyquery==1.4.3
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.5.4
requests==2.28.1
requests-html==0.10.0
rlp==3.0.0
selenium==4.3.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
surrogates==1.0.2
tblib==1.7.0
toolz==0.12.0
tornado==4.5.3
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
u-msgpack-python==2.7.1
urllib3==1.26.12
urllib3-secure-extra==0.1.0
w3lib==2.1.1
web3==6.2.0
websockets==10.3
Werkzeug==2.1.2
wincertstore==0.2
WsgiDAV==4.0.1
wsproto==1.1.0
XlsxWriter==3.0.8
xvfbwrapper==0.2.9
yarl==1.7.2
zipp==3.8.0
